Extract blocknote theme detection `useOpTheme` hook
